requirements.txt
# TerminatorBot - Global Prediction Market Exploitation System
# Python 3.12+

# ── Prediction Market APIs ────────────────────────────
pmxt>=2.0.0
betfairlightweight[speed]>=2.20.0
# kalshi-python>=2.1.0        # Fallback if pmxt Kalshi has issues
# py-clob-client>=0.34.0      # Fallback if pmxt Polymarket has issues

# ── Machine Learning ─────────────────────────────────
xgboost>=2.0.0
scikit-learn>=1.4.0
textblob>=0.18.0

# ── Fuzzy Matching ───────────────────────────────────
rapidfuzz>=3.6.0

# ── Data & HTTP ──────────────────────────────────────
requests>=2.31.0
aiohttp>=3.9.0
pandas>=2.1.0
numpy>=1.26.0

# ── Config & Auth ────────────────────────────────────
python-dotenv>=1.0.0
cryptography>=41.0.0

# ── Terminal UI ──────────────────────────────────────
colorama>=0.4.6
tabulate>=0.9.0

# ── Testing ──────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
